pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer Flask Application"
authors = [
    {name = "Hexlet", email = "support@hexlet.io"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "validators>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[tool.ruff]
target-version = "py38"
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ARG", "SIM", "TCH", "TID", "Q", "RSE", "RET", "SLF", "SLOT", "PIE", "LOG", "PTH", "EM", "NPY", "AIR", "PERF", "FURB", "TRY", "FLY", "A", "COM", "C90", "DTZ", "ISC", "ICN", "G", "INP", "T20", "PYI", "PT"]
ignore = ["E501", "W503", "F401"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] 